Move code specific to Google login to separate fragment
