+++ WordPress/src/main/java/org/wordpress/android/modules/AppComponent.java
@@ -26,6 +26,7 @@ import org.wordpress.android.ui.accounts.login.Login2FaFragment;
 import org.wordpress.android.ui.accounts.login.LoginEmailFragment;
 import org.wordpress.android.ui.accounts.login.LoginEmailPasswordFragment;
 import org.wordpress.android.ui.accounts.login.LoginEpilogueFragment;
+import org.wordpress.android.ui.accounts.login.LoginGoogleFragment;
 import org.wordpress.android.ui.accounts.login.LoginMagicLinkRequestFragment;
 import org.wordpress.android.ui.accounts.login.LoginSiteAddressFragment;
 import org.wordpress.android.ui.accounts.login.LoginSiteAddressHelpDialogFragment;
@@ -135,6 +136,7 @@ public interface AppComponent {
     void inject(LoginActivity object);
     void inject(LoginMagicLinkRequestFragment object);
     void inject(LoginEmailFragment object);
+    void inject(LoginGoogleFragment object);
     void inject(LoginEmailPasswordFragment object);
     void inject(Login2FaFragment object);
     void inject(LoginSiteAddressFragment object);
